Avoid clash between choria and task parameters
In #13, an `--environment` parameter was added to choria to select which
environment should be used to download a task files.  However, this
conflicts with tasks which expect and `environment` parameter and break
them.

Rename the environment selection option `-E` to avoid this issue: as a
task name is always lowercase, this should not introduce any conflict.