(#19) Fix FreeBSD dependency on execution_wrapper
execution_wrapper(1) is a wrapper command part of the pxp-agent package,
which depends on the cpp-pcp-client.  These two packages have been
archived upstream, and the later does not build with current versions of
boost.  Due to the way FreeBSD packages are built and shipped, this
make it impossible for FreeBSD to use Bolt tasks with Choria.

To improve the situation, the rubygem-choria-mcorpc-support FreeBSD
package now include its own execution_wrapper(1) and makes this
dependency unnecessary.